requirements.txt
crewai
langchain_community
python-dotenv
exa_py
crewai_tools
markdown2
typing
requests
pdfminer
streamlit
